blood-test-analyzer (3)/requirements.txt
click>=8.1.8
crewai==0.130.0 
fastapi==0.110.3
google-ai-generativelanguage==0.6.4
google-api-core>=2.15.0,<3.0.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.53.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
Jinja2==3.1.4
jsonschema==4.22.0
langchain-core==0.1.52
langsmith==0.1.67
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.22.0
openai>=1.68.2
pandas==2.2.2
pillow==10.3.0
protobuf>=3.19.5,!=3.20.0,!=3.20.1,!=4.21.0,!=4.21.1,!=4.21.2,!=4.21.3,!=4.21.4,!=4.21.5,<5.0.0
pydantic>=2.4.2,<3.0.0
python-dotenv>=1.0.0
uvicorn>=0.29.0
python-multipart>=0.0.9
langchain-community>=0.0.38
pypdf>=4.0.0
reportlab>=4.0.0
requests>=2.31.0
